refactor(test): split SystemTest into focused UI-helper mixins (Closes #831)
SystemTest had grown to ~540 lines mixing terminal, tabs, splits, SSH,
password prompt, monitoring, SFTP, and settings concerns — past the
project's ~500-line / single-responsibility guideline, with every further
ported E2E suite pushing it further over.

Split it into a thin base (per-suite app lifecycle + the `wait` polling
primitive + the `delay4user` watch-along hook) plus focused, composable
`*Ui` mixins that a suite opts into ahead of `SystemTest` in its bases:
TerminalUi, TabsUi, LayoutUi, SidebarUi, ConnectionsUi, PasswordPromptUi,
SshUi, MonitoringUi, SftpUi, SettingsUi. `ui.py` becomes a `ui/` package
(one module per concern); the name->element store lookups stay free
functions so they remain unit-testable without an app. A shared
HarnessMixin declares the borrowed `driver`/`wait` for type-checkers
without shadowing the real base members at runtime.

Each suite now declares exactly what it drives, e.g.
`class TestX(ConnectionsUi, TabsUi, SystemTest)`. Pure refactor — no
behavior change; helper bodies are byte-identical, only relocated.

Adds test_mixins.py (machinery group, no app) asserting the mixins
compose into one MRO and that `driver`/`wait` are never shadowed by a
mixin — locking in the split's key invariant.
