+++ tests/system/termihub_harness/__init__.py
@@ -28,6 +28,15 @@ from .orchestrator import AgentInstance, AppInstance, agent_binary_path, app_bin
 from .systemtest import SystemTest, unique_name
 from .ui import (
     ConnectionsUi,
+    LayoutUi,
+    MonitoringUi,
+    PasswordPromptUi,
+    SettingsUi,
+    SftpUi,
+    SidebarUi,
+    SshUi,
+    TabsUi,
+    TerminalUi,
     connection_item_testid,
     connections,
     find_connection,
@@ -46,6 +55,15 @@ __all__ = [
     "agent_binary_path",
     "SystemTest",
     "ConnectionsUi",
+    "TerminalUi",
+    "TabsUi",
+    "LayoutUi",
+    "SidebarUi",
+    "PasswordPromptUi",
+    "SshUi",
+    "MonitoringUi",
+    "SftpUi",
+    "SettingsUi",
     "connections",
     "find_connection",
     "folders",
