+++ tests/system/tests/test_ssh_password_extended.py
@@ -9,11 +9,16 @@ from __future__ import annotations
 import pytest
 
 from termihub_harness import (
-    SSH_KEY_PATH,
+    ConnectionsUi,
+    PasswordPromptUi,
     SSH_KEYS_PORT,
+    SSH_KEY_PATH,
     SSH_PASSWORD_PORT,
     SSH_USERNAME,
+    SidebarUi,
     SystemTest,
+    TabsUi,
+    TerminalUi,
     unique_name,
 )
 
@@ -23,7 +28,7 @@ HOST = "127.0.0.1"
 
 
 @pytest.mark.usefixtures("ssh_fixtures")
-class TestSshPasswordExtended(SystemTest):
+class TestSshPasswordExtended(TerminalUi, TabsUi, SidebarUi, ConnectionsUi, PasswordPromptUi, SystemTest):
     def test_key_auth_shows_no_password_dialog(self):
         name = unique_name("ssh-key-nopass")
         self.create_ssh_connection(
